# app/api/v1/openapi.json
{"openapi":"3.1.0","info":{"title":"Hesabi API","description":"A comprehensive Personal Financial Management and Budget Tracking API.","version":"1.0.0"},"paths":{"/api/v1/categories/{category_id}":{"get":{"tags":["Categories"],"summary":"Get single category","description":"Retrieve specific category details and its budget goal by ID.","operationId":"get_category_api_v1_categories__category_id__get","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Categories"],"summary":"Edit a category","description":"Partially update a category name or its budget goal. Only provided fields are updated.","operationId":"edit_category_api_v1_categories__category_id__patch","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Category Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Categories"],"summary":"Delete a category","description":"Permanently delete a category and its associated budget from the database.","operationId":"delete_category_api_v1_categories__category_id__delete","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Category Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/categories/":{"get":{"tags":["Categories"],"summary":"List of all categories","description":"Fetch a list of all expense and budget categories.","operationId":"get_all_categories_api_v1_categories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoryResponse"},"type":"array","title":"Response Get All Categories Api V1 Categories  Get"}}}}}},"post":{"tags":["Categories"],"summary":"Create category","description":"Create a new category alongside its assigned budget goal.","operationId":"create_category_api_v1_categories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transactions/{transaction_id}":{"get":{"tags":["Transactions"],"summary":"Get single transaction","description":"Retrieve detailed information of a specific transaction by its unique ID.","operationId":"get_transaction_api_v1_transactions__transaction_id__get","parameters":[{"name":"transaction_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Transaction Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Transactions"],"summary":"Edit a transaction","description":"Partially update an existing transaction. Only fields supplied in the request body will be modified.","operationId":"edit_transaction_api_v1_transactions__transaction_id__patch","parameters":[{"name":"transaction_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Transaction Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Transactions"],"summary":"Delete a transaction","description":"Permanently remove a transaction from the database by its unique ID.","operationId":"delete_transaction_api_v1_transactions__transaction_id__delete","parameters":[{"name":"transaction_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Transaction Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transactions/":{"get":{"tags":["Transactions"],"summary":"List of all transactions","description":"Fetch an ordered list of all recorded income and expense transactions.","operationId":"get_all_transaction_api_v1_transactions__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TransactionResponse"},"type":"array","title":"Response Get All Transaction Api V1 Transactions  Get"}}}}}},"post":{"tags":["Transactions"],"summary":"Create a transaction","description":"Record a new income or expense transaction linked optionally to a category.","operationId":"create_transaction_api_v1_transactions__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/compute/summary":{"get":{"tags":["Compute"],"summary":"Get overall financial summary","description":"Aggregate all transactions to calculate total income, total expenses, and net balance.","operationId":"get_summary_api_v1_compute_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FinancialSummaryResponse"}}}}}}},"/api/v1/compute/categories":{"get":{"tags":["Compute"],"summary":"Get all categories","description":"Calculate budget utilization, total expenses, and remaining limits across all categories.","operationId":"get_all_categories_api_v1_compute_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoryBalanceResponse"},"type":"array","title":"Response Get All Categories Api V1 Compute Categories Get"}}}}}}},"/api/v1/compute/{category_id}":{"get":{"tags":["Compute"],"summary":"Get a single category","description":"Compute the spending progress and remaining budget for a specific category by ID.","operationId":"get_category_api_v1_compute__category_id__get","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Category Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryBalanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Auth"],"summary":"Register a new user","description":"Register a new user account with email, name, and strong password.","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Auth"],"summary":"Refresh","operationId":"refresh_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefreshRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefreshResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["Users"],"summary":"Get current user profile","description":"Retrieve profile details of the authenticated and active user via Bearer token.","operationId":"get_me_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"CategoryBalanceResponse":{"properties":{"name":{"type":"string","maxLength":20,"minLength":1,"title":"Name","examples":["fast food"]},"budget_goal":{"type":"integer","exclusiveMinimum":0.0,"title":"Budget Goal","examples":[5000000]},"spent":{"type":"integer","minimum":0.0,"title":"Spent","examples":[100000]},"remaining":{"type":"integer","title":"Remaining","examples":[-200000]}},"type":"object","required":["name","budget_goal","spent","remaining"],"title":"CategoryBalanceResponse","example":{"budget_goal":5000000,"name":"Fast Food","remaining":-200000,"spent":5200000}},"CategoryCreate":{"properties":{"name":{"type":"string","maxLength":20,"minLength":1,"title":"Name","examples":["Fast Food"]},"budget_goal":{"type":"integer","exclusiveMinimum":0.0,"title":"Budget Goal","examples":[500000]}},"type":"object","required":["name","budget_goal"],"title":"CategoryCreate","example":{"budget_goal":8000000,"name":"Supermarket"}},"CategoryResponse":{"properties":{"name":{"type":"string","maxLength":20,"minLength":1,"title":"Name","examples":["Fast Food"]},"budget_goal":{"type":"integer","exclusiveMinimum":0.0,"title":"Budget Goal","examples":[500000]},"id":{"type":"integer","title":"Id","examples":[2]}},"type":"object","required":["name","budget_goal","id"],"title":"CategoryResponse","example":{"budget_goal":8000000,"id":1,"name":"Supermarket"}},"CategoryUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":20,"minLength":1},{"type":"null"}],"title":"Name","examples":["Sport"]},"budget_goal":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Budget Goal","examples":[700000]}},"type":"object","title":"CategoryUpdate"},"FinancialSummaryResponse":{"properties":{"income":{"type":"integer","title":"Income","examples":[400000]},"expense":{"type":"integer","title":"Expense","examples":[120000]},"total":{"type":"integer","title":"Total","examples":[280000]},"currency":{"type":"string","title":"Currency","default":"TOMAN","examples":["TOMAN"]}},"type":"object","required":["income","expense","total"],"title":"FinancialSummaryResponse","example":{"currency":"TOMAN","expense":120000,"income":400000,"total":280000}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"Token"},"TokenRefreshRequest":{"properties":{"refresh_token":{"type":"string","minLength":100,"pattern":"^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"TokenRefreshRequest"},"TokenRefreshResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenRefreshResponse"},"TransactionCreate":{"properties":{"amount":{"type":"integer","exclusiveMinimum":0.0,"title":"Amount","examples":[150000]},"type":{"type":"string","enum":["income","expense"],"title":"Type","examples":["income"]},"date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Date","examples":["2026-08-30T10:30:00Z"]},"note":{"anyOf":[{"type":"string","maxLength":225},{"type":"null"}],"title":"Note","examples":["this is a note"]},"category_id":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Category Id","examples":[5]}},"type":"object","required":["amount","type"],"title":"TransactionCreate","example":{"amount":150000,"category_id":5,"date":"2026-08-30T10:30:00Z","note":"Freelance project payment","type":"income"}},"TransactionResponse":{"properties":{"amount":{"type":"integer","exclusiveMinimum":0.0,"title":"Amount","examples":[150000]},"type":{"type":"string","enum":["income","expense"],"title":"Type","examples":["income"]},"date":{"type":"string","format":"date-time","title":"Date","examples":["2026-08-30T10:30:00Z"]},"note":{"anyOf":[{"type":"string","maxLength":225},{"type":"null"}],"title":"Note","examples":["this is a note"]},"category_id":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Category Id","examples":[5]},"id":{"type":"integer","exclusiveMinimum":0.0,"title":"Id","examples":[2]},"currency":{"type":"string","title":"Currency","default":"TOMAN","examples":["TOMAN"]}},"type":"object","required":["amount","type","date","id"],"title":"TransactionResponse","example":{"amount":150000,"category_id":5,"currency":"TOMAN","date":"2026-08-30T10:30:00Z","id":2,"note":"Freelance project payment","type":"income"}},"TransactionUpdate":{"properties":{"amount":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Amount","examples":[130000]},"type":{"anyOf":[{"type":"string","enum":["income","expense"]},{"type":"null"}],"title":"Type","examples":["expense"]},"date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Date","examples":["2026-08-30T10:30:00Z"]},"note":{"anyOf":[{"type":"string","maxLength":150},{"type":"null"}],"title":"Note","examples":["Updated note"]},"category_id":{"anyOf":[{"type":"integer","exclusiveMinimum":0.0},{"type":"null"}],"title":"Category Id","examples":[5]}},"type":"object","title":"TransactionUpdate","example":{"amount":130000,"note":"Updated payment note"}},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"name":{"anyOf":[{"type":"string","maxLength":150,"minLength":1},{"type":"null"}],"title":"Name","examples":[]},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"UserRegister"},"UserResponse":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"name":{"anyOf":[{"type":"string","maxLength":150,"minLength":1},{"type":"null"}],"title":"Name","examples":[]},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","id","is_active","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}},"tags":[{"name":"Categories","description":"Manage financial categories and set monthly budget goals."},{"name":"Transactions","description":"Record and manage income and expense transactions."},{"name":"Compute","description":"Calculate financial rollups, overall summaries, and category balance,"},{"name":"Auth","description":"User registration, authentication, and JWT token management."},{"name":"Users","description":"Operation on authenticated user profiles,"}]}